Divide linker output parameters by rms
Flux related parameters (max, min, sum) returned by the linker module are now divided by the global cube rms (which may be = 1 in case a weights cube has been applied). These parameters are now called SNRmax, SNRmin and SNRsum. They are used for the reliability calculation. These parameters used to be replaced by the values returned by the parameterisation module. This is not the case any more since the former are in units of (loacal) RMS, the latter in BUNIT.

To make these changes I have had to move the definition of the catalogue header to before the reliability calculation.