Change nrbins
I have set the initial value of nrbins to scale with the cube size. It now is 1e-5 times the number of voxels in the cube. Testing on a few cubes suggest that this is a reasonable value. Note that "nrbins" appears in the definition of GetRMS but its value is now overwritten by this new default. We should remove nrbins from the function definition if we are happy with the new default.