Dilation of neighbouring objects
When dilating object "i", we set flux=0 for all voxels belonging to objects!="i" in the dilation subcube. This ensures that their flux is not included in the flux growth curve. Only the binary mask of object "i" is dilated. Other objects are left unchanged. Object "i" cannot grow "on top" of other objects.