RMS in subcube to speed up code
Calculate the RMS using the entire input cube ("rms") as well as on a subcube (zoomx=10,zoomy=10,zoomz=10) ("rms_sample"). When smoothing with a kernel, calculate the RMS on a subcube only (zoomx=10,zoomy=10,zoomz=10) and use the ratio to "rms_sample" to scale "rms". Speeds up S+C by a factor of ~1.5 for a 1GB cube. Could be made more sophisticated.