Improve subcube checks for subCubeMode="world"
The checks that the requested subcube is within the input cube were insufficient when using subCubeMode="world". This should now be improved. If the requested subcube has zero overlap with the input cube along an axis, SoFiA gives an error message -- informing the user of which axis causes the problem -- and fails with exit code 1.