Measure rms in subcube
I've implemented a new default behaviour of S+C where the noise is estimated in the central 100x100x100 subcube. This does not affect the noise calculations in other parts of the code, just S+C. For the moment the size of the subcube along each axis is hardcoded to 100, but we could make this a SoFiA "advanced" parameter accessible through the parameter file.