Major edit
This major edit enables SoFiA to automatically find the optimal smoothing kernel necessary to derive the reliability of positive sources as a function of position in parameter space. The reliability is calculated by convolving the distribution of positive and negative sources with the kernel. The basic method consists of finding a kernel large enough such that the positive and negative density fields satisfy the "Skellam criterion" outlines in Sec. 4 of Serra, Jurek & Floer (2012).