Configurable GUI With this update, the GUI now is configurable by the user. Settings will automatically be stored and retrieved from a hidden configuration file, named .SoFiA, in the user's home directory. All settings can be modified by hand, but using the configuration options offered through the GUI is the recommended way of changing settings.
Currently available settings include: (1) display of toolbar, (2) display of pipeline message window, (3) docking position of pipeline message window, (4) automatic saving of current session on shutdown, (5) width of the main window and (6) height of the main window.

When exiting from SoFiA, the current state of the main window will automatically be saved and restored at the beginning of the next session.