Changed globalrms and related steps
The calculation of globalrms now happens within the reliability section of sofia_pipeline.py . Source parameters related to the reliability calculation (i.e., snr_sum, snr_max, snr_min, n_pos, n_neg) are only included in the debugging output catalogue. They are not included in the regular output catalogue. This solves the inconsistencies we were having with processing filtered cubes, where 1) the globalrms noise as well as the parameters snr_max, snr_min, snr_sum are not defined and cannot be calculated.

Note that the lines producing the debugging masks and moments are commented out as they are broken. This needs to be fixed.

I have tested these changes with as many different SoFiA settings as I could manage. As usual please test further and report problems.