Change in kernel calculation and output
The initial automatic kernel for deriving the positive and negative density fields in parameter space can now be either the covariance or the variance matrix. The former is useful to smooth the distribution of points with kernels aligned according to the correlation between parameters, potentially better preserving the edges of the distribution. A new parameter is introduced ("usecov") to allow the user to decide whether to use the full covariance matrix or just the variance matrix.

The other change is a new diagnostic plot produced as part  of the kernel optimization. The new plot shows the median (N-P)/sqrt(N+P) value as a function of kernel size. This is useful to establish whether the adopted kernel size convergence criterion makes sense (typically, convergence happens when this median value is above a negative number close to zero, the current default being -0.2).

This commit includes other changes to the structure of the code, but these do not change the actual calculations done by the code or the final result.