Bug fix and change in calculation of R
Fixed a bug (using sigma instead of variance at the beginning of the code).

Also changed the way we estimate the number of positives and negatives from the density fields. We used to take the value of the PDF at the desired location. Now we take the integral of the PDF within the convolution kernel centred at the desired location.