changed kernel growth
Before this change the variance of the smoothing kernel used to calculate the reliability was iteratively grown by a given factor along each axis. This meant that the N-d kernel volume would grow faster when calculating the reliability in a space with more dimensions (i.e., when using more source parameters). This was giving problems (e.g., overshooting kernel size) when working with >3 parameters.

I have now changed the code such that the growth factor is evelvated to the power of 1/(N-1) where N is the number of parameters being used. Therefore, in the new code the kernel volume grows always by the same factor regardless of the number of parameters being used.