import subcube of mask
The import of a mask is now consistent with the import of data and weights when a subcube is requested. This was not the case, and when reading in a subcube of the data and weights the full mask was still imported. The resulting mismatch between mask and cube shape was causing in an incorrect list of detection when using subcubes. Strangely, it was not causing the code to crash...