fixed cubelets-subcube incompatibility
When working on subcubes the Linker returns objects' XYZ coordinates and bounding boxes in the subcube frame. These are then transformed to the original, larger cube frame, such that the final catalogue bounding boxes can be used on the original cube. This transformation was happening before the extraction of cubelets. This was a problem because cubelets are extracted from the subcube and, therefore, require XYZ objects' coordinates and bounding boxes in the subcube frame, not in the original, larger cube frame. I have now moved the transformation to after the extraction of cubelets. The transformation still happens before the WCS module, which is the desired  behaviour because WCS coordinates are calculated using the original, larger cube header.