Added regridding of cubelets
Cubelets are now regridded before making the moment images. This produces moment images consistent with those obtained from the parent cube. A small difference remains. It is due to the fact that regridding is performed on both mask (converted to float binary) and cube. The regridded mask includes voxel values different from just 0 and 1 (e.g., <0). We then set everything below threshold = abs(min(regridded_mask)) to 0, and everything above to 1. We then proceed as in the no-regrid case. When processing the parent cube a single threshold is applied to the parent regridded mask and, therefore, to all objects. When processing individual cublets each has its own threshold. Small differences in the threshold values cause small differences in the regridded mask once converted to binary and, therefore, small differences in the moment images. We may want to address this at a later stage.