mask subcube selection
If a user requests to extract a subcube of the input cube and provides an input mask, SoFiA now uses either the full mask or the subcube of the mask. The choice depends on which of the two matches the size of the selected subcube. If none does, SoFiA returns an error and stops. Additionally, SoFiA checks that the CRVAL's of the mask are the same as the ones of the cube. This is a simple test that the WCS's are likely the same.