cleaned up code and fixed mom maps' units
I have improved the way cubelets.py deals with different 3rd-axis types/units.

The new code can deal with a 3rd axis in frequency (as long as the CTYPE3 value includes the string "FREQ" -- case insensitive) and velocity (for CTYPE3 equal to VOPT, VRAD, VELO or FELO -- case insensitive). For these two cases, if no units are given (i.e., no CUNIT3 is present in the header) the code assumes Hz and m/s, respectively, and mom0, 1 and 2 images are scaled accordingly. If the units are m/s or km/s  they are converted to km/s and Hz, respectively, and mom0, 1 and 2 images are scaled accordingly. For all other 3rd-axis types and/or units the code makes no conversion: it multiplies the moment-0 image by CDELT3 and makes use of CUNIT3 if given, else a "std_unit_"+CTYPE3 is reported in the header.

Note that writemoment2.pyx should be made consistent with cubelets.py .