Implemented user-defined kernel scaling
I've added a new parameter called scaleKernel. This allows the user to adopt a kernel whose shape is that of the (co-)variance matrix, scaled by the factor scaleKernel. This functionality is available at the GUI. So the kernel size can now be set in one of the following 3 ways:
1) fully user defined (not in the GUI)
2) auto shape equal to the (co-)variance matrix, and scale given by the user (in the GUI except for the use of the variance)
3) auto shape equal to the (co-)variance matrix, and auto scale through the algorithm in this module (in the GUI except for the use of the variance)