removed "import astropy.io.fits as astropy"
    import astropy.io.fits as astropy
was a very bad idea. Not only because astropy is an module, which we actively use in SoFiA, but because we are doing things like 
    from astropy import wcs
at other places in the module. This can be a become a bit confusing and lead to errors. We should check if the rest of the code contains similar things.