Reliability with no negative sources
This change addresses https://github.com/SoFiA-Admin/SoFiA/issues/113 (situation 1; errror due to the absence of negative voxels in the cube). With this change, if a user requests the reliability calculation but no negative sources were detected SoFiA returns an error message + a few suggestions, and exits.

This change addresses the absence of negative voxels because such absence necessarily implies the absence of negative sources.

Note that SoFiA now exits before attempting the measurement of the global rms, which was the reason for crashing. This measurement was done to normalise the source's flux parameters in preparation of the reliability calculation.

In addition, this change also addresses situations in which no negative sources are detected even in the presence of negative voxels, e.g., because of a very high source finding threshold. This situation was already taken care of within addrel.py but it is now caught earlier, before addrel.EstimateRel is called within sofia_pipeline.py.