Two-pass std with 5 sigma cut
The std rms calculation is very fast so a default two-pass strategy seems like a good idea. Not sure whether we want to let users revert back to a one-pass strategy. For the moment I have not included that flexibility.

The 5 sigma cut is not very aggressive so the rms won't change in the absence of outliers. Based on some tests I think we could even use a 4 sigma cut.