Introduce option to provide indexed mask to generate output products without filtering or merging.
Source parameters that would usually be returned by the linker module are
generated from the input mask. This allows for the execution of subsequent
pipeline modules, that require these parameters.

Note: Changes needed to be made to the dilation function in parametrisation.py
to account for cases, in which the sources are not labelled consecutively.