Fix bug in moment zero generation for 2D input files
This commit should resolve issue #155, where SoFiA crashes on 2D images because it is expecting file header keywords for a third dimension. The header is now checked for the 'CTYPE3' keyword. If it is missing moment zero images are generated without taking a third axis into account and the generation of moment one images is skipped. There are still a few problems with output products for 2D input files that should be fixed (see issue #156).